requirements.txt
APScheduler==3.10.1
b64uuid==0.1
boto3==1.26.92
botocore==1.29.92
cryptography==39.0.2
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.14